Throw error if number is bigger than MAX size
The function toWords throws "RangeError: Maximum call stack size exceeded" exception
if the number is greater than MAX size.

The changes from this patch introduces a check to verify if the number can be transformed into words.
If the number is greater than MAX size, it throws an explicit exception to pinpoint the problem.